Improved the chessAI to an elo of at least ~900

+++ Chess Game/board.h
@@ -28,7 +28,7 @@ namespace BoardStructure {
 	extern BoardState initBoard;
 	extern Magic::color currMoveColor;
 
-	int getBoardValue();
+	double getBoardValue();
 	Piece::Base* find(sf::Vector2i);
 	void initTextures();
 	void init();
